+++ components/Navbar/Navbar.module.scss
@@ -11,31 +11,3 @@
     color: red;
   }
 }
-
-.Modal {
-  display: flex;
-  height: 70vh;
-  &__employer,
-  &__candidate {
-    width: 70%;
-    border-radius: 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    &__text {
-      text-align: center;
-    }
-  }
-  &__candidate {
-    background-color: $green;
-  }
-  &__candidate:hover {
-    background-color: $lightgreen;
-  }
-  &__employer {
-    background-color: $orange;
-  }
-  &__employer:hover {
-    background-color: $lightorange;
-  }
-}
